+++ FMSystems.WeatherForecast.Api/FMSystems.WeatherForecast.Api.csproj
@@ -6,14 +6,12 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="5.0.8" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.8" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\FMSystems.WeatherForecast.Infrastructure\FMSystems.WeatherForecast.Infrastructure.csproj" />
     <ProjectReference Include="..\FMSystems.WeatherForecast.Web\FMSystems.WeatherForecast.Web.csproj" />
     <ProjectReference Include="..\FMSystems.WeatherForecast.Domain\FMSystems.WeatherForecast.Domain.csproj" />
-    <ProjectReference Include="..\FMSystems.WeatherForecast.Application\FMSystems.WeatherForecast.Services.csproj" />
   </ItemGroup>
