--- FMSystems.WeatherForecast.Application/FMSystems.WeatherForecast.Services.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\FMSystems.WeatherForecast.Domain\FMSystems.WeatherForecast.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
