--- FMSystems.WeatherForecast.Application/IWeatherForecastService.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace FMSystems.WeatherForecast.Services
-{
-    public interface IWeatherForecastService
-    {
-        IEnumerable<WeatherForecast.Domain.WeatherForecast> GetForecasts();
-    }
-}
