+++ FMSystems.WeatherForecast.Domain/Entity/BaseEntity.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace FMSystems.WeatherForecast.Domain
+namespace FMSystems.WeatherForecast.Domain.Entity
 {
     public class BaseEntity
     {
