+++ FMSystems.WeatherForecast.Domain/Service/IWeatherForecastService.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace FMSystems.WeatherForecast.Domain.Service
+{
+    public interface IWeatherForecastService
+    {
+        IEnumerable<WeatherForecast.Domain.Entity.WeatherForecast> GetForecasts();
+    }
+}
