+++ FMSystems.WeatherForecast.Domain/Service/WeatherForecastService.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using FMSystems.WeatherForecast.Domain;
+using FMSystems.WeatherForecast.Domain.Service;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -11,10 +13,10 @@ namespace FMSystems.WeatherForecast.Services
             "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
         };
 
-        public IEnumerable<FMSystems.WeatherForecast.Domain.WeatherForecast> GetForecasts()
+        public IEnumerable<FMSystems.WeatherForecast.Domain.Entity.WeatherForecast> GetForecasts()
         {
             var rng = new Random();
-            return Enumerable.Range(1, 5).Select(index => new FMSystems.WeatherForecast.Domain.WeatherForecast
+            return Enumerable.Range(1, 5).Select(index => new FMSystems.WeatherForecast.Domain.Entity.WeatherForecast
             {
                 Date = DateTime.Now.AddDays(index),
                 TemperatureC = rng.Next(-20, 55),
