+++ FMSystems.WeatherForecast.Infrastructure/Db/Context/IWeatherForecastDbContext.cs
@@ -1,7 +1,8 @@
 ﻿using FMSystems.WeatherForecast.Domain;
+using FMSystems.WeatherForecast.Domain.Entity;
 using Microsoft.EntityFrameworkCore;
 
-namespace FMSystems.WeatherForecast.Infrastructure.DBContexts
+namespace FMSystems.WeatherForecast.Infrastructure.Db.Context
 {
     public interface IWeatherForecastDbContext
     {
