+++ FMSystems.WeatherForecast.Infrastructure/Db/Context/WeatherForecastDbContext.cs
@@ -1,6 +1,8 @@
 ﻿
 using FMSystems.WeatherForecast.Domain;
-using FMSystems.WeatherForecast.Infrastructure.SeedData;
+using FMSystems.WeatherForecast.Domain.Entity;
+using FMSystems.WeatherForecast.Infrastructure.Db.Context;
+using FMSystems.WeatherForecast.Infrastructure.Db.SeedData;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -8,7 +10,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace FMSystems.WeatherForecast.Infrastructure.DBContexts
+namespace FMSystems.WeatherForecast.Infrastructure.DBContext
 {
     public class WeatherForecastDbContext : DbContext, IWeatherForecastDbContext
     {
