+++ FMSystems.WeatherForecast.Infrastructure/FMSystems.WeatherForecast.Infrastructure.csproj
@@ -11,4 +11,9 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="5.0.8" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Api\Repository\" />
+    <Folder Include="Api\" />
+  </ItemGroup>
 </Project>
