+++ FMSystems.WeatherForecast.Web/Pages/FetchData.razor
@@ -1,5 +1,5 @@
 ﻿@page "/fetchdata"
-@using FMSystems.WeatherForecast.Domain
+@using FMSystems.WeatherForecast.Domain.Entity
 @inject HttpClient Http
 
 <h1>Weather forecast</h1>
