+++ FMSystems.WeatherForecast.sln
@@ -2,8 +2,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.0.0
 MinimumVisualStudioVersion = 16.0.0.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FMSystems.WeatherForecast.Services", "FMSystems.WeatherForecast.Application\FMSystems.WeatherForecast.Services.csproj", "{5F4146C3-463A-453A-8635-B81308E63516}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{11321F96-7329-4F4D-B45D-9287627F2093}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
@@ -35,18 +33,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5F4146C3-463A-453A-8635-B81308E63516}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Debug|x64.Build.0 = Debug|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Debug|x86.Build.0 = Debug|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Release|x64.ActiveCfg = Release|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Release|x64.Build.0 = Release|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Release|x86.ActiveCfg = Release|Any CPU
-		{5F4146C3-463A-453A-8635-B81308E63516}.Release|x86.Build.0 = Release|Any CPU
 		{F026821E-DA29-4C36-B565-ECF043A54087}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F026821E-DA29-4C36-B565-ECF043A54087}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F026821E-DA29-4C36-B565-ECF043A54087}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -100,7 +86,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{5F4146C3-463A-453A-8635-B81308E63516} = {B38A8620-96F4-47DB-81AC-CB0F7D7B09D0}
 		{F026821E-DA29-4C36-B565-ECF043A54087} = {64400B20-008D-4832-9982-4C5B4FAE1303}
 		{9AE7E397-799B-4ECD-BB00-9EE462D4E0E2} = {A4FB2807-8F43-40E6-9FD1-EC17C0E022E8}
 		{1BC7FC11-A57B-4514-B495-42787AF4FEC1} = {B38A8620-96F4-47DB-81AC-CB0F7D7B09D0}
